+++ examples/complete/variables.tf
@@ -47,7 +47,7 @@ variable "region" {
 }
 
 variable "subnets" {
-  type = "string"
+  type = "list"
 }
 
 variable "vpc_id" {